finding user check added to user service

+++ laravel-app/app/Domains/Users/Services/UserService.php
@@ -18,13 +18,17 @@ class UserService
 
     public function getAUser(User $user)
     {
+        if(!$user){
+            return "user not found";
+        }
         return $user;
     }
 
     public function updateUser($request, User $user)
     {
-        echo $user;
-
+        if(!$user){
+            return "user not found";
+        }
         $user->update([
             'name' => $request -> name,
             'password' => $request -> password,
@@ -36,6 +40,9 @@ class UserService
 
     public function deleteUser(User $user)
     {
+        if(!$user){
+            return "user not found";
+        }
         return $user->delete();
     }
 }
